# packages/client/spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Ownware Gateway — core wire contract
  version: 0.29.0
  description: >
    Drive a Ownware agent from any language: start a run, stream its
    events, answer permission prompts. Bearer auth applies to every
    route except /health whenever the gateway has auth enabled (always
    true on non-loopback binds; token at <dataDir>/gateway-token).
servers:
  - url: http://localhost:3011
    description: Default local gateway (ownware serve prints yours)
security:
  - bearer: []
paths:
...
  /api/v1/source-data-views/{dataViewId}/access-grants:
    post:
      operationId: createDataViewQueryGrant
      summary: Owner-only admission of one exact Data View observation fence.
      description: >
        Authentication must be enabled and the bearer must be the install owner.
        The request names explicit current field identities and one bounded row window;
        the Gateway atomically derives and stores the exact stable row identities. The
        fixed grant is source_data_view / source_data_views.query / observe. It grants
        no manifest, source-content, search, write, filter, sort, aggregation, SQL or
        expression authority. Exact idempotency retries return the same minimal receipt.
      parameters:
        - { in: path, name: dataViewId, required: true, schema: { type: string, format: uuid } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDataViewQueryGrantInput' }
      responses:
        '201':
          description: Minimal immutable mutation receipt; no cells, row identities or policy detail.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content: { application/json: { schema: { $ref: '#/components/schemas/AccessGrantMutationReceipt' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/source-resources/{resourceId}/access-grants:
    post:
      operationId: createAccessGrant
      summary: Owner-only creation of one bounded protected-source read or search grant.
      description: >
        Authentication must be enabled and the bearer must be the install owner.
        Workspace/profile scope and immutable source lineage are derived by the
        Gateway. An exact UUID idempotency retry returns only the same minimal receipt.
      parameters:
        - { in: path, name: resourceId, required: true, schema: { type: string, format: uuid } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccessGrantInput' }
      responses:
        '201':
          description: Minimal immutable mutation receipt; no subject, consent, policy, or storage detail.
          content: { application/json: { schema: { $ref: '#/components/schemas/AccessGrantMutationReceipt' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/access-grants:
    get:
      operationId: listAccessGrants
      summary: Owner-only bounded list of current grant revisions and lifecycle truth.
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: cursor, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Current grant heads ordered by opaque identity.
          content: { application/json: { schema: { $ref: '#/components/schemas/AccessGrantList' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/access-grants/{grantId}:
    get:
      operationId: getAccessGrant
      summary: Owner-only read of one current immutable grant revision.
      parameters:
        - { in: path, name: grantId, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Current immutable grant revision.
          content: { application/json: { schema: { $ref: '#/components/schemas/AccessGrant' } } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  responses:
    ErrorResponse:
      description: Safe typed error. Raw internal/provider detail is not returned.
      headers:
        X-Ownware-Correlation-Id:
          schema: { type: string, format: uuid }
          description: Opaque server-generated support correlation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    CreateDataViewQueryGrantInput:
      type: object
      required: [subjectId, purpose, channel, consent, fieldIds, rowOffset, rowCount, ttlSeconds]
      properties:
        subjectId: { type: string, minLength: 1, maxLength: 128 }
        purpose: { type: string, minLength: 1, maxLength: 64 }
        channel: { type: [string, 'null'], minLength: 1, maxLength: 64 }
        consent: { $ref: '#/components/schemas/AccessConsent' }
        fieldIds:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { type: string, pattern: '^field\.[0-9a-f]{32}$' }
        rowOffset:
          type: integer
          minimum: 0
          maximum: 99999
          description: First row ordinal in the exact grant window.
        rowCount:
          type: integer
          minimum: 1
          maximum: 256
          description: Exact row count; a window extending beyond the current manifest rejects rather than clipping.
        ttlSeconds: { type: integer, minimum: 60, maximum: 2592000 }
      additionalProperties: false
...
    AccessConsent:
      oneOf:
        - type: object
          required: [state]
          properties:
            state: { type: string, const: not_required }
          additionalProperties: false
        - type: object
          required: [state, evidenceId]
          properties:
            state: { type: string, const: recorded }
            evidenceId:
              type: string
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
          additionalProperties: false
    CreateAccessGrantInput:
      type: object
      required: [subjectId, purpose, channel, consent, ttlSeconds]
      properties:
        operation: { type: string, enum: [source_content.read, source_content.search] }
        subjectId: { type: string, minLength: 1, maxLength: 128 }
        purpose: { type: string, minLength: 1, maxLength: 64 }
        channel: { type: [string, 'null'], minLength: 1, maxLength: 64 }
        consent: { $ref: '#/components/schemas/AccessConsent' }
        ttlSeconds: { type: integer, minimum: 60, maximum: 2592000 }
      additionalProperties: false
    AccessGrantMutationReceipt:
      type: object
      required: [grantId, revision, mutation, acceptedAt]
      properties:
        grantId: { type: string, format: uuid }
        revision: { type: integer, minimum: 1 }
        mutation: { type: string, enum: [created, revoked] }
        acceptedAt: { type: integer, minimum: 0 }
      additionalProperties: false
    AccessGrantProjection:
      type: object
      required: [grantId, revision, state, workspaceId, profileId, subjectId, purpose, channel, resourceKind, resourceId, operation, fieldScope, rowScope, consent, autonomyCeiling, effectiveAt, expiresAt, issuedBy, revisionCreatedAt, revokedAt]
      properties:
        grantId: { type: string, format: uuid }
        revision: { type: integer, minimum: 1 }
        state: { type: string, enum: [active, revoked] }
        workspaceId: { type: string }
        profileId: { type: string }
        subjectId: { type: string }
        purpose: { type: string }
        channel: { type: [string, 'null'] }
        resourceKind: { type: string, enum: [source_resource, source_data_view] }
        resourceId: { type: string, format: uuid }
        operation: { type: string, enum: [source_content.read, source_content.search, source_data_views.query] }
        fieldScope:
          oneOf:
            - { $ref: '#/components/schemas/AllAccessScope' }
            - { $ref: '#/components/schemas/FieldListAccessScope' }
        rowScope:
          oneOf:
            - { $ref: '#/components/schemas/AllAccessScope' }
            - { $ref: '#/components/schemas/RowListAccessScope' }
        consent: { $ref: '#/components/schemas/AccessConsent' }
        autonomyCeiling: { type: string, const: observe }
        effectiveAt: { type: integer }
        expiresAt: { type: integer }
        issuedBy: { type: string, const: install_owner }
        revisionCreatedAt: { type: integer }
        revokedAt: { type: [integer, 'null'] }
    AccessGrant:
      allOf:
        - { $ref: '#/components/schemas/AccessGrantProjection' }
      unevaluatedProperties: false
    AllAccessScope:
      type: object
      required: [mode]
      properties:
        mode: { type: string, const: all }
      additionalProperties: false
    FieldListAccessScope:
      type: object
      required: [mode, ids]
      properties:
        mode: { type: string, const: list }
        ids:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { type: string, pattern: '^field\.[0-9a-f]{32}$' }
      additionalProperties: false
    RowListAccessScope:
      type: object
      required: [mode, ids]
      properties:
        mode: { type: string, const: list }
        ids:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { type: string, pattern: '^row\.[0-9a-f]{32}$' }
      additionalProperties: false
    CurrentAccessGrant:
      allOf:
        - { $ref: '#/components/schemas/AccessGrantProjection' }
        - type: object
          required: [lifecycle]
          properties:
            lifecycle: { type: string, enum: [scheduled, effective, expired, revoked] }
      unevaluatedProperties: false
    AccessGrantList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CurrentAccessGrant' }
        nextCursor: { type: [string, 'null'], format: uuid }
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, message, category, correlationId]
      properties:
        error: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        message: { type: string, maxLength: 500 }
        category: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        correlationId: { type: string, format: uuid }
        retryAfter:
          type: number
          minimum: 0
          description: Optional retry delay in seconds.
        actualRevision:
          type: integer
          minimum: 1
          description: Safe actual source revision, present only for `source_upload_refresh_conflict`.
        actualCurrentVersionId:
          type: [string, 'null']
          format: uuid
          description: Safe actual current source version, present only for `source_upload_refresh_conflict`.
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Gateway session token. Enabled automatically on any non-loopback
        bind; persisted at <dataDir>/gateway-token (0600).